ticket_validator/requirements.txt
django==2.1.7
djangorestframework==3.9
djangorestframework_simplejwt==3.3
drf-yasg==1.14
django-cors-headers==2.4.0
gunicorn==19.9.0
coreapi
pyyaml
mysqlclient
